ai-agent/requirements.txt
google-adk[extensions]==1.0.0
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-dotenv
openai
